refactor: extract Snyk from google-cloud package to root
Snyk is executed when @verdaccio/google-cloud package is published, using prepublish script, but other packages could be published before due to topological order of Lerna publish lifecycle (like @verdaccio/babel-preset or @verdaccio/eslint-config). Also, if the package is removed from monorepo, Snyk will never be executed.
With this commit we try to extract Snyk configuration to the root package, so it will not depend on the execution of a package, making it generic for all. Lastly, as we don't want to publish packages from local computers (it's done by CI environment) but we update versions manually, the command execution has been moved to preversion script, to perform locally before making a stub commit.