fix(verdaccio-htpasswd): supply old password to crypt3 on change
To generate the proper comparison string, crypt3 needs the salt from the
original entry. This previously worked only due to the constant salt.
The verifyPassword function already does this correctly.

Add a test that verifies that the password change works and a
different result is returned when a new salt is generated.